infoCore {CORElearn}	R Documentation

Description of certain CORElearn parameters

Description

Depending on parameter what the function prints some information on CORElearn, for example codes of available classification (or regression) attribute evaluation heuristics. For more complete description of the parameters see helpCore.

Usage

infoCore(what=c("attrEval","attrEvalReg"))

Arguments

what
Selects the info to be printed.

Details

Depending on the parameter what the function some information on CORElearn.

"attrEval"

Prints codes of all available classification attribute evaluation heuristics. These codes can be used as parameters for attribute evaluation methods in learning. It is internally used for validation of parameters. For more complete information see attrEval.

"attrEvalReg"

prints codes of all available regression attribute evaluation heuristics. These codes can be used as parameters for attribute evaluation methods in learning. It is internally used for validation of parameters. For more complete information see attrEval.

Value

For what="attrEval" or "attrEvalReg" function returns vector of codes for all implemented classification or regression attribute evaluation heuristics, respectively.

Author(s)

Marko Robnik-Sikonja

See Also

attrEval, helpCore, CoreModel.

Examples

estClass <- infoCore(what="attrEval")
print(estClass)
infoCore(what="attrEvalReg")

[Package CORElearn version 1.48.0 Index]