noEqualRows {CORElearn}	R Documentation

Number of equal rows in two data sets

Description

Counts number of equal rows in two data sets. The two data sets shall have equal number of columns.

Usage

 noEqualRows(data1, data2, tolerance=1e-5, countOnce=TRUE)

Arguments

data1
The first data set.

data2
The second data set.

tolerance
Tolerated difference between two rows.

countOnce
Shall each equal row in data1 be counted just once, or number of rows it is equal to in data2.

Details

Rows are compared using column-wise comparisons. The sum of differences up to a given tolerance are tolerated.

Value

Integer value giving the count of equal instances.

Author(s)

Marko Robnik-Sikonja

See Also

CORElearn.

Examples

# uses two randomly generated data sets
set.seed(12345)
d1 <- classDataGen(100)
d2 <- classDataGen(100)
noEqualRows(d1, d2, tolerance=1e-4)

[Package CORElearn version 1.48.0 Index]